refactor: use project reference for backend
