+++ apps/web/deprecated/index-copy.tsx
@@ -1,7 +1,7 @@
 import { useActor, useService } from '@xstate/react'
 import { Button, Modal, Table } from 'antd'
 import React, { useContext } from 'react'
-// import { withApollo } from '@codelab/ddd/shared/infrastructure'
+// import { withApollo } from '@codelab/ddd/backend/infrastructure'
 import { UserSignupButton } from '@codelab/ddd/modules/users-stories'
 import { collectionToTable } from '@codelab/shared/factory'
 import { EventNameApp } from '@codelab/state/app'
