--- libs/ddd/backend/framework/.eslintrc.js
@@ -1,19 +1,0 @@
-module.exports = {
-  extends: '../../../../.eslintrc.js',
-  root: true,
-  settings: {
-    'import/resolver': {
-      typescript: {
-        project: 'tsconfig.eslint.json',
-      },
-    },
-  },
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ['tsconfig.eslint.json'],
-  },
-  rules: {
-    'react/static-property-placement': 'off',
-    'no-prototype-builtins': 'off',
-  },
-}
