+++ libs/ddd/backend/infrastructure/tsconfig.json
@@ -1,5 +1,4 @@
 {
-  "extends": "../../../../tsconfig.base.json",
   "include": [
     "**/*.ts",
     "**/*.tsx",
