+++ libs/ddd/backend/tsconfig.eslint.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.base.json",
+  "extends": "./tsconfig.json",
   "include": [
     "**/*.ts",
     "**/*.tsx",
