+++ libs/ddd/modules/users/core/application/useCases/createUser/CreateUserUseCase.spec.ts
@@ -12,7 +12,7 @@ import { UserUseCaseDto } from '../UserUseCaseDto'
 import {
   TestDatabaseModule,
   TypeOrmUser,
-} from '@codelab/ddd/shared/infrastructure'
+} from '@codelab/ddd/backend/infrastructure'
 
 describe('CreateUserUseCase', () => {
   let userModule: any
