+++ libs/ddd/modules/users/core/application/useCases/createUser/CreateUserUseCase.ts
@@ -1,6 +1,6 @@
 import { CreateUserRequest } from './CreateUserRequest'
 import { CreateUserResponse } from './CreateUserResponse'
-import { TransactionalUseCase } from '@codelab/ddd/shared/core'
+import { TransactionalUseCase } from '@codelab/ddd/backend/core'
 
 export type CreateUserUseCase = TransactionalUseCase<
   CreateUserRequest,
